Fix commas inside "lr".
Here we resort to a hack. Officially, "lr" takes a single body
argument that is a VContent.  But, the way we parse arguments,
parsing of a VContent will stop when we hit a comma. So, everything
after the first comma of the argument of "lr" was being ignored.
We handle this by changing the type to a list of VContents, and then
concatenating them with commas between.

This gives good results for e.g. `lr(a,b,c)`.