Fix precedence for functions.
`1(x)` and `!(x)` should not be parsed as functions.

Closes #55.

Note that we still don't match typst's behavior for `f_"!"(x)`.
For us this works just like `f_!(x)`, but for typst we get
a function in the subscript for the former but not the latter.
Fixing this would require some changes in the types.