Show rule changes.
ShowRule now has an extra parameter for a unique identifier.
This allows us to prevent double application of show rules,
while allowing distinct rules with the same selector.
[API change]

In applying show rules, we no longer recurse into an element's
fields, as this caused double application of show rules in nested
contexts. (See #63.)

However, this is not a complete fix because there are some tests
that still fail.