Fix incorrect parsing of underscores in code (#82)
Typst does not allow `_` (a single underscore) as an identifier and
rather lexes it in code as a special token `Underscore`.
<https://github.com/typst/typst/blob/630fbbec415c5d0dd854b62d6039b11dd10c4961/crates/typst-syntax/src/lexer.rs#L810>

The consequence is that

```typst
#let m=1
_#m._
```

is valid and parses as `#emph` with content of the value of `m` followed
by a period.

On the other hand,

```typst
#{
  _
}
```

and

```typst
#let d = (_: 1)
```

fail with `expected expression, found underscore`.

However, this library currently completely fails parsing the first
example, and the second example fails in evaluation with `Identifier "_"
not found`, and the third example evaluates without errors.

This commit fixes the above cases and aligns the behavior closer to
Typst.

- Disallows a single underscore as an identifier in `pBasicExpr` parser.
- In `pExpr`, parses a single underscore (when not in a pattern) as
  `Expr Underscore`. The evaluation of this expression then prints the
expected error message.
- Adds 3 tests for underscore handling in code for the above 3 examples.