Stop line comments from eating the end of line character
This is important for correctly issuing parbreaks and spaces that
follow line comments and a preparation for simpler handling of indented
blocks.
    
The accepted output `test/typ/regression/issue90-comment-00.out`
illustrates the difference: a parbreak is correctly inserted in

```typ
par one//
     
par two
```

Unfortunately, this tweak changes the output of ~500 tests since most
tests contain comments (but brings it closer to Typst's whitespace
handling).

The changed tests are added in the child commit for easier testing.